feat: add schemas for BestPracticeImportEntry, ChecklistItem, and ChecklistResult; update Pipeline and Skill schemas
