+++ packages/schemas/src/SkillSchema.ts
@@ -7,8 +7,8 @@ export const SkillSchema = z.object({
   description: z.string(),
   category: z.string(),
   tags: z.array(z.string()),
-  createdAt: z.number(),
-  updatedAt: z.number(),
+  createdAt: z.coerce.date(),
+  updatedAt: z.coerce.date(),
 });
 
 export type Skill = z.infer<typeof SkillSchema>;
