+++ packages/schemas/src/index.ts
@@ -1,6 +1,8 @@
 export * from "./BestPracticeSchema";
+export * from "./BestPracticeImportEntrySchema";
 export * from "./BestPracticeImportSchema";
-export * from "./ChecklistSchema";
+export * from "./ChecklistItemSchema";
+export * from "./ChecklistResultSchema";
 export * from "./CodeSnippetSchema";
 export * from "./GithubProjectSchema";
 export * from "./JobSchema";
